# docs/openapi.yaml
openapi: 3.1.0
info:
  title: XLS-66 Lending Protocol Reference App
  description: |
    HTTP API for the XRPL Lending Protocol (XLS-66) + Single Asset Vault
    (XLS-65) reference implementation. Every route that mutates on-chain
    state maps 1:1 to an XRPL transaction and re-queries the ledger before
    returning.

    ## Conventions

    - **Auth** — handled by Auth0 Universal Login. Clients obtain a session
      by visiting `/auth/login` (browser only) — there is no programmatic
      login endpoint. The session is stored in an httpOnly cookie `xls66-auth`.
      Every `/api/*` route is gated by `src/middleware.ts` and returns
      `401 Unauthorized` without a valid cookie.
    - **Amounts** — drops (integer string) for XRP; decimal string for IOU
      and MPT. See the README for the scaling conventions.
    - **Rates** — UI-facing rates are % or basis points; on-ledger they are
      1/10 bps. The API accepts the UI form and scales at the boundary.
    - **Content type** — `application/json` for all request/response bodies.
  version: 1.0.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: XLS-66 reference app
servers:
  - url: http://localhost:3000
    description: Local dev server
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xls66-auth
      description: httpOnly cookie set by `POST /api/session`. Value is the Session document's MongoDB ObjectId.
  parameters:
    LoanIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, description: "Hash256 of the Loan or LoanBroker ledger entry" }
...
  schemas:
...
    Loan:
      description: MongoDB cache row for a loan. Cached terms + latest synced state.
      type: object
      required: [_id, sessionId, loanId, loanBrokerId, borrowerAddress]
      properties:
        _id: { type: string }
        sessionId: { type: string }
        loanId: { type: string }
        loanBrokerId: { type: string }
        borrowerAddress: { type: string }
        principalRequested: { type: string }
        interestRate: { type: integer, description: "Stored as basis points." }
        paymentTotal: { type: integer }
        paymentInterval: { type: integer, description: "Seconds" }
        gracePeriod: { type: integer, description: "Seconds" }
        originationFee: { type: string }
        serviceFee: { type: string }
        paymentsRemaining: { type: integer }
        principalOutstanding: { type: string }
        status: { type: string, enum: [active, repaid, defaulted, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BrokerOptions:
      description: All rates in 1/10 bps. Range 0-10_000 for ManagementFeeRate, 0-100_000 for cover rates.
      type: object
      properties:
        managementFeeRate: { type: integer, minimum: 0, maximum: 10000 }
        debtMaximum: { type: integer, minimum: 0 }
        coverRateMinimum: { type: integer, minimum: 0, maximum: 100000 }
        coverRateLiquidation: { type: integer, minimum: 0, maximum: 100000 }
    TxResult:
      description: Pass-through of the rippled `tx` submission result. Only the hash is canonical; everything else is for debugging.
      type: object
      additionalProperties: true
      properties:
        hash: { type: string }
        ledger_index: { type: integer }
        meta:
          type: object
          additionalProperties: true
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid `xls66-auth` cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Referenced entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Unhandled error or rippled failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
paths:
...
  /api/broker:
    post:
      tags: [Broker]
      summary: Register broker (LoanBrokerSet + optional cover deposit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vaultId]
              properties:
                vaultId: { type: string }
                brokerOptions: { $ref: "#/components/schemas/BrokerOptions" }
                coverAmountDrops:
                  type: string
                  description: First-loss cover to deposit immediately. Drops for XRP, decimal for IOU/MPT.
      responses:
        "200":
          description: Broker registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  loanBrokerId: { type: string }
                  result: { $ref: "#/components/schemas/TxResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [Broker]
      summary: Delete broker (LoanBrokerDelete)
      description: Caller must have previously withdrawn any remaining cover.
      responses:
        "200":
          description: Broker deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/loan/{id}:
    get:
      tags: [Loan]
      summary: Loan or LoanBroker detail
      description: |
        The same route serves both `Loan` and `LoanBroker` entries — pass the
        ledger-entry id regardless of type. Amount fields are unscaled for
        MPT vaults.
      parameters:
        - $ref: "#/components/parameters/LoanIdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  loan: { $ref: "#/components/schemas/Loan" }
                  onLedger:
                    type: object
                    additionalProperties: true
        "500": { $ref: "#/components/responses/ServerError" }
  /api/loan/repay:
    post:
      tags: [Loan]
      summary: Make a payment (LoanPay)
      description: |
        Mode mapping:

        | `mode`         | xrpl.js flag           | Amount source                       |
        | -------------- | ---------------------- | ----------------------------------- |
        | `"full"`       | `tfLoanFullPayment`    | Server-computed, XLS-66 §A-3.2.4    |
        | `"late"`       | `tfLoanLatePayment`    | Server-computed, XLS-66 §A-3.2.2    |
        | `"overpayment"`| `tfLoanOverpayment`    | `amountDrops`                       |
        | `"regular"` / omitted | —               | `amountDrops` (ledger caps)         |

        `amountDrops` is required for `overpayment`/`regular`/`custom` paths.
        It is ignored for `full` and `late` — the server re-derives the
        exact amount from the latest validated ledger close time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loanId]
              properties:
                loanId: { type: string }
                amountDrops:
                  type: string
                  description: "Drops for XRP, decimal for IOU/MPT. Omit for full/late."
                mode:
                  type: string
                  enum: [full, late, overpayment, regular]
      responses:
        "200":
          description: Payment submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
                  amount:
                    type: string
                    description: The amount the server actually submitted. Useful for full/late to show the real debit.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/loan/default:
    post:
      tags: [Loan]
      summary: Default or close a loan
      description: |
        With `action: "default"` (or omitted) the server submits `LoanManage`
        with the `tfLoanDefault` flag. With `action: "close"` it submits
        `LoanDelete` (only valid on fully repaid or defaulted loans).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loanId]
              properties:
                loanId: { type: string }
                action:
                  type: string
                  enum: [default, close]
                  default: default
      responses:
        "200":
          description: Loan defaulted or deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { $ref: "#/components/schemas/TxResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }